JAM-0309 MemeMaker Meme maker 指令

+++ plugins/meme_maker/src/main/scala/jam/plugins/meme_maker/engine/MemeAPIRequest.scala
@@ -8,7 +8,7 @@ import upickle.default._
  * Author: sinar
  * 2020/11/18 22:25
  */
-case class MemeAPIRequest(id: Int, fillings: Map[String, String])
+case class MemeAPIRequest(id: Long, fillings: Map[String, String])
 object MemeAPIRequest {
   implicit val rw: ReadWriter[MemeAPIRequest] = macroRW
 
@@ -17,7 +17,7 @@ object MemeAPIRequest {
    *
    * @return 请求体
    */
-  def apply(id: Int, sentences: List[String]): String = write(
+  def apply(id: Long, sentences: List[String]): String = write(
     new MemeAPIRequest(id, sentences.zipWithIndex.map {
       case (sentence, idx) => s"sentence$idx" -> sentence
     }.toMap)
